% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlpa_likelihood.R
\name{mlpa_likelihood}
\alias{mlpa_likelihood}
\title{calculate likelhood of MLPA}
\usage{
mlpa_likelihood()
}
\description{
\code{mlpa_likelihood} returns the negative log likelihood
of the MLPA data
}

