% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_demon.R
\name{prep_demon}
\alias{prep_demon}
\title{Prepare objects required for Laplaces Demon}
\usage{
prep_demon(demondat, pos_vars, scale_numerics = T)
}
\arguments{
\item{demondat}{data to be used}

\item{pos_vars}{possible variables to include in the model}

\item{scale_numerics}{scale T or F to center and scale variables}
}
\description{
\code{prep_demon} takes data and options and
returns a list with components required by Laplaces Demon
}

